+++ src/Core/DB.php
@@ -711,7 +711,7 @@ class DB {
 			name: $settingName,
 			description: "DB version of {$fileBase}",
 			mode: 'noedit',
-			type: 'text',
+			type: (is_int($version) || preg_match('/^\d+$/', $version)) ? 'timestamp' : 'text',
 			value: "0"
 		);