+++ src/Core/TextSettingHandler.php
@@ -38,7 +38,7 @@ class TextSettingHandler extends SettingHandler {
 	public function displayValue(string $sender): string {
 		$displayValue = parent::displayValue($sender);
 		if ($displayValue === "<highlight><end>") {
-			$displayValue = "<highlight>&lt;empty&gt;<end>";
+			$displayValue = "<grey>&lt;empty&gt;<end>";
 		}
 		return $displayValue;
 	}