+++ src/Modules/SKILLS_MODULE/BuffPerksController.php
@@ -75,8 +75,8 @@ class BuffPerksController extends ModuleInstance {
 	public LoggerWrapper $logger;
 
 	/** DB version of perks */
-	#[NCA\Setting\Text(mode: 'noedit')]
-	public string $perksDBVersion = "0";
+	#[NCA\Setting\Timestamp(mode: 'noedit')]
+	public int $perksDBVersion = 0;
 
 	/** @var Collection<Perk> */
 	public Collection $perks;
@@ -93,10 +93,7 @@ class BuffPerksController extends ModuleInstance {
 		}
 		$path = __DIR__ . "/perks.csv";
 		$mtime = @filemtime($path);
-		$dbVersion = 0;
-		if ($this->settingManager->exists("perks_db_version")) {
-			$dbVersion = (int)$this->settingManager->get("perks_db_version");
-		}
+		$dbVersion = $this->perksDBVersion;
 		$perkInfo = $this->getPerkInfo();
 		$this->perks = new Collection($perkInfo);
 		$empty = !$this->db->table("perk")->exists();